Auto_Licence_Clean/requirements.txt
google-cloud-bigquery==3.25.0
google-auth==2.29.0
requests==2.32.3

# Dev / tests uniquement
pytest==8.3.4
